% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{detect_studies}
\alias{detect_studies}
\title{Detect study columns (est/stderr) in a meta result data frame}
\usage{
detect_studies(df)
}
\description{
Detect study columns (est/stderr) in a meta result data frame
}
