% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{forest_plot}
\alias{forest_plot}
\title{Make a forest plot and save to file}
\usage{
forest_plot(
  df_long,
  y_levels,
  outFile,
  title = NULL,
  xlab = "Effect",
  xlim_num = NULL,
  het_y = NULL,
  width = 10,
  height = 7,
  dpi = 300,
  show_meta = FALSE,
  df_meta = NULL
)
}
\description{
Make a forest plot and save to file
}
