% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateInfo.R
\name{parse_snp_ids}
\alias{parse_snp_ids}
\title{Parse SNP IDs of the form "chr1:1119172:G:A"}
\usage{
parse_snp_ids(snp_ids)
}
\arguments{
\item{snp_ids}{Character vector of SNP IDs.}
}
\value{
Data frame with CHR, SNP, POS, A1, A2.
}
\description{
Parse SNP IDs of the form "chr1:1119172:G:A"
}
