% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot_manhattan}
\alias{plot_manhattan}
\title{Draw a Manhattan plot and save to JPEG}
\usage{
plot_manhattan(
  df,
  outFile,
  title = NULL,
  pCut = 1e-05,
  onlySig = FALSE,
  width = 11,
  height = 4.5,
  dpi = 300
)
}
\description{
Draw a Manhattan plot and save to JPEG
}
