% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot_qq}
\alias{plot_qq}
\title{Draw a QQ plot and save to file}
\usage{
plot_qq(df, outFile, title = NULL, width = 6, height = 6, dpi = 300)
}
\arguments{
\item{df}{Data frame containing a p-value column named `pval`.}

\item{outFile}{Output path (png/jpg/etc).}

\item{title}{Optional plot title.}

\item{width, height, dpi}{Device parameters.}
}
\description{
Draw a QQ plot and save to file
}
