% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{read_result_file}
\alias{read_result_file}
\title{Read result file and harmonize columns (works for step3 meta or single-study step2)}
\usage{
read_result_file(path, sep = "\\t", study_id = NULL)
}
\description{
Read result file and harmonize columns (works for step3 meta or single-study step2)
}
