% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{to_long_study}
\alias{to_long_study}
\title{Convert wide study columns to long format for forest plotting}
\usage{
to_long_study(df_rows, studies, y_col, ciMult = 1.96, studyLabels = NULL)
}
\description{
Convert wide study columns to long format for forest plotting
}
