% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateInfo.R
\name{feature_info}
\alias{feature_info}
\title{Compute feature-level prevalence and average proportion}
\usage{
feature_info(abd_file)
}
\arguments{
\item{abd_file}{Path to abundance table (tab-separated, first column sample ID).}
}
\value{
A data frame with FeatureID, Prevalence, AvgProportion.
}
\description{
Compute feature-level prevalence and average proportion
}
