% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateInfo.R
\name{run_info}
\alias{run_info}
\title{Run all step0 info generation and write outputs}
\usage{
run_info(geno_file, abd_file, output_snp, output_feature, output_seqdepth)
}
\arguments{
\item{geno_file}{PLINK prefix for genotype files.}

\item{abd_file}{Path to abundance table.}

\item{output_snp}{Output path for SNP info.}

\item{output_feature}{Output path for feature info.}

\item{output_seqdepth}{Output path for sequencing depth info.}
}
\value{
Invisibly returns a list of output paths.
}
\description{
Run all step0 info generation and write outputs
}
