% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateInfo.R
\name{seqdepth_info}
\alias{seqdepth_info}
\title{Compute sequencing depth per sample}
\usage{
seqdepth_info(abd_file)
}
\arguments{
\item{abd_file}{Path to abundance table (tab-separated, first column sample ID).}
}
\value{
A data frame with SampleID and SeqDepth.
}
\description{
Compute sequencing depth per sample
}
