% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateInfo.R
\name{snp_info}
\alias{snp_info}
\title{Compute SNP counts and allele frequency from PLINK files}
\usage{
snp_info(geno_prefix)
}
\arguments{
\item{geno_prefix}{PLINK prefix (without .bed/.bim/.fam).}
}
\value{
Data frame with CHR, SNP, POS, A1, A2, N, AF.
}
\description{
Compute SNP counts and allele frequency from PLINK files
}
